+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Property `TranslationMap.len_per_source`.
+
 ### Changes
 - Improve `PandasIO` performance (~4x faster for large datasets).
 - Typing updates (notably numpy `2.0`).